move combo.js => combo.mjs
